refactor editor slice
