+++ app/javascript/slices/editor.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-param-reassign */
 import { createSlice } from '@reduxjs/toolkit';
-import { useDispatch } from 'react-redux';
 
 const slice = createSlice({
   name: 'editor',
@@ -13,27 +12,9 @@ const slice = createSlice({
     updateCode(state, { payload }) {
       state.code = payload;
     },
-    runCodeSuccess(state, { payload }) {
-      state.stateOfRuningCode = payload;
-    },
   },
 });
 
-const { runCodeSuccess } = slice.actions;
-
-export const useRunCode = () => {
-  const dispatch = useDispatch();
-
-  const fakeResponse = 'Code has been running successful!';
-  const runCode = () => new Promise((resolve) => {
-    setTimeout(() => resolve(dispatch(runCodeSuccess(fakeResponse))), 1000);
-  });
-
-  return { runCode };
-};
-
-export const {
-  updateCode,
-} = slice.actions;
+export const { actions } = slice;
 
 export default slice.reducer;
