+++ app/javascript/slices/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
-import editor from './editor.js';
+import editor, { actions as editorActions } from './editor.js';
 import terminal from './terminal.js';
 
 
@@ -8,3 +8,7 @@ export default combineReducers({
   editor,
   terminal,
 });
+
+export const actions = {
+  ...editorActions,
+};
